+++ settings/course_formats.php
@@ -37,16 +37,6 @@ if ($ADMIN->fulltree) {
         get_string('coursepagesidebarinfooterenabledsection', 'theme_adaptable'),
         format_text(get_string('coursepagesidebarinfooterenabledsectiondesc', 'theme_adaptable'), FORMAT_MARKDOWN)));
 
-    // Commonly used activities and resources.
-    require_once($CFG->dirroot . '/course/lib.php');
-    $name = 'theme_adaptable/commonlyusedar';
-    $title = get_string('commonlyusedar', 'theme_adaptable');
-    $description = get_string('commonlyusedardesc', 'theme_adaptable',
-        '\''.implode(', ', array_keys(get_module_types_names())).'\'');
-    $default = '';
-    $setting = new admin_setting_configtextarea($name, $title, $description, $default);
-    $page->add($setting);
-
     $name = 'theme_adaptable/coursepagesidebarinfooterenabled';
     $title = get_string('coursepagesidebarinfooterenabled', 'theme_adaptable');
     $description = get_string('coursepagesidebarinfooterenableddesc', 'theme_adaptable');